KAFKA/docker-compose/alert-consumer/requirements.txt
flask==3.0.0
flask-socketio==5.3.5
kafka-python==2.0.2
python-socketio==5.10.0
